addLogBootstrap method inside the Logger class registers the ILogger interface by loading a platform-specific concrete class (using Class.forname()) to public log field. Actual registration-instantiation happens inside static{} blocks of IOIOAndroidApplicationHelper and IOIOPcApplicationHelper
