+++ software/applications/HelloIOIO/proguard/seeds.txt
@@ -100,12 +100,11 @@ ioio.lib.impl.SequencerImpl$RemoteState: ioio.lib.impl.SequencerImpl$RemoteState
 ioio.lib.impl.SequencerImpl$RemoteState: ioio.lib.impl.SequencerImpl$RemoteState valueOf(java.lang.String)
 ioio.lib.impl.SocketIOIOConnectionBootstrap: SocketIOIOConnectionBootstrap()
 ioio.lib.spi.Logger: Logger()
-ioio.lib.spi.Logger$ApplicationType: ioio.lib.spi.Logger$ApplicationType[] values()
-ioio.lib.spi.Logger$ApplicationType: ioio.lib.spi.Logger$ApplicationType valueOf(java.lang.String)
 ioio.lib.util.AbstractIOIOActivity: AbstractIOIOActivity()
 ioio.lib.util.BaseIOIOLooper: BaseIOIOLooper()
 ioio.lib.util.IOIOConnectionManager$Thread: IOIOConnectionManager$Thread()
 ioio.lib.util.IOIOConnectionRegistry: IOIOConnectionRegistry()
 ioio.lib.util.IOIOLooperAlt: IOIOLooperAlt()
+ioio.lib.util.android.ConcreteLog: ConcreteLog()
 ioio.lib.util.android.IOIOActivity: IOIOActivity()
 ioio.lib.util.android.IOIOService: IOIOService()
